pyproject.toml
[project]
name = "cache-redis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "google-cloud-bigquery>=3.30.0",
    "redis>=5.2.0",
    "uvicorn>=0.34.0",
]
